Change filterCoaches computed property
It will no longer show all the coaches, but the ones who are filtered. Save coaches in constant coaches and call the filter built-in method to first check if the specific expertise is even checked and then see if the coach's area includes that specific expertise. Return true if matched so that the referred coach is displayed.