Check if the inputs are valid using the validateForm method
First sets the validateFrom to valid and then goes through the inputs. If one of the inputs is empty, the referred input and the form is considered invalid. For the rate it is checked whether the it is null or the value is less than zero. For the areas it is checked whether the length is equal to zero.