Use a global navigation guard
It is used to prevent users who are not authenticated from entering certain pages. If the path the user tries to enter requires authentication but is not authorized, the user will be redirected to the auth page. If the path does not require authentication but the user is already authorized, the user will be redirected to the coaches page. In other conditions, navigation is allowed,